0935.py
# Project 935. Multi-modal Sentiment Analysis - MODERNIZED VERSION

# This is the original simple implementation that has been modernized into a full research-ready project.
# The modernized version includes:
# - Modern deep learning architecture (DistilBERT + CNN-LSTM)
# - Advanced fusion strategies (late fusion, cross-attention)
# - Comprehensive evaluation metrics
# - Interactive Streamlit demo
# - Production-ready code structure

# ORIGINAL SIMPLE IMPLEMENTATION (for reference):

import librosa
import numpy as np
from textblob import TextBlob
import torch
import librosa.display
 
# Step 1: Sentiment Analysis from Text using TextBlob
def text_sentiment(text):
    sentiment = TextBlob(text).sentiment.polarity
    if sentiment > 0:
        return "Positive"
    elif sentiment < 0:
        return "Negative"
    else:
        return "Neutral"
 
# Example text input
text_input = "I love this product, it's amazing!"
text_sentiment_result = text_sentiment(text_input)
print(f"Text Sentiment: {text_sentiment_result}")
 
# Step 2: Sentiment Analysis from Audio using librosa
def audio_sentiment(audio_file):
    y, sr = librosa.load(audio_file)
    
    # Extract pitch (fundamental frequency) and tempo features from audio
    pitch, _ = librosa.core.piptrack(y=y, sr=sr)
    tempo, _ = librosa.beat.beat_track(y=y, sr=sr)
 
    # Analyze the features to determine sentiment (simplified logic)
    avg_pitch = np.mean(pitch[pitch > 0])  # average pitch (higher is often more positive)
    avg_tempo = tempo  # tempo (higher tempo may indicate excitement)
 
    if avg_pitch > 200 and avg_tempo > 120:
        return "Positive"
    elif avg_pitch < 100 and avg_tempo < 100:
        return "Negative"
    else:
        return "Neutral"
 
# Example audio input (replace with a valid audio file path)
audio_input = "example_audio.wav"
audio_sentiment_result = audio_sentiment(audio_input)
print(f"Audio Sentiment: {audio_sentiment_result}")
 
# Combine both text and audio sentiment results
final_sentiment = "Overall Sentiment: " + ("Positive" if text_sentiment_result == "Positive" and audio_sentiment_result == "Positive" else "Negative" if text_sentiment_result == "Negative" or audio_sentiment_result == "Negative" else "Neutral")
print(final_sentiment)

# MODERNIZED VERSION FEATURES:
# ===========================
# 
# 1. MODERN ARCHITECTURE:
#    - Text Encoder: DistilBERT (pre-trained transformer)
#    - Audio Encoder: CNN + Bidirectional LSTM
#    - Fusion: Late fusion, cross-attention, contrastive learning
#    - Classifier: Multi-layer perceptron with dropout
#
# 2. ADVANCED TRAINING:
#    - Mixed precision training (AMP)
#    - Gradient clipping and learning rate scheduling
#    - Early stopping and checkpointing
#    - Data augmentation (audio noise, time masking)
#
# 3. COMPREHENSIVE EVALUATION:
#    - Accuracy, Macro-F1, Weighted-F1
#    - Confusion matrix and classification report
#    - Calibration metrics (ECE)
#    - Per-class performance analysis
#
# 4. INTERACTIVE DEMO:
#    - Streamlit-based real-time analysis
#    - Audio file upload and visualization
#    - Confidence scores and probability distributions
#    - Waveform and spectrogram plots
#
# 5. PRODUCTION READY:
#    - Clean, typed code with docstrings
#    - Comprehensive test suite
#    - Configuration management (YAML)
#    - Logging and visualization utilities
#    - Device management (CUDA/MPS/CPU)
#
# TO USE THE MODERNIZED VERSION:
# =============================
# 
# 1. Run the interactive demo:
#    streamlit run demo/streamlit_app.py
#
# 2. Train a model:
#    python scripts/train.py --config configs/config.yaml
#
# 3. Run simple demo:
#    python demo_simple.py
#
# 4. Run tests:
#    pytest tests/
#
# The modernized version transforms this simple script into a research-ready
# multimodal sentiment analysis system with state-of-the-art performance.

